+++ projects/shell/src/environments/environment.dev.ts
@@ -1,6 +1,6 @@
 export const environment = {
   production: true,
-  manifestPath: '/assets/mf.manifest.dev.json?v=1785316816170',
+  manifestPath: '/assets/mf.manifest.dev.json?v=1785318749622',
   mfe: {
     admissions: 'https://admissions.dev-kjusys.kristujayanti.edu.in/',
     core: 'https://core.dev-kjusys.kristujayanti.edu.in/',